serve/requirements.txt
flask
pillow
torch
torchvision
numpy
cv2
